Address most PR comments (except extra-copy)
